# Conflicts:
#	src/Craft.php

+++ src/elements/GlobalSet.php
@@ -40,6 +40,14 @@ class GlobalSet extends Element
     /**
      * @inheritdoc
      */
+    public static function pluralDisplayName(): string
+    {
+        return Craft::t('app', 'Global Sets');
+    }
+
+    /**
+     * @inheritdoc
+     */
     public static function refHandle()
     {
         return 'globalset';
@@ -93,7 +101,7 @@ class GlobalSet extends Element
      */
     public function __toString(): string
     {
-        return (string)$this->name;
+        return (string)$this->name ?: static::class;
     }
 
     /**
@@ -149,11 +157,11 @@ class GlobalSet extends Element
      */
     public function getCpEditUrl()
     {
-        if (Craft::$app->getIsMultiSite() && $this->siteId != Craft::$app->getSites()->getCurrentSite()->id) {
-            return UrlHelper::cpUrl('globals/'.$this->getSite()->handle.'/'.$this->handle);
+        if (Craft::$app->getIsMultiSite()) {
+            return UrlHelper::cpUrl('globals/' . $this->getSite()->handle . '/' . $this->handle);
         }
 
-        return UrlHelper::cpUrl('globals/'.$this->handle);
+        return UrlHelper::cpUrl('globals/' . $this->handle);
     }
 
     // Events
@@ -164,10 +172,27 @@ class GlobalSet extends Element
      */
     public function beforeDelete(): bool
     {
+        if (!parent::beforeDelete()) {
+            return false;
+        }
+
         if (($fieldLayout = $this->getFieldLayout()) !== null) {
             Craft::$app->getFields()->deleteLayout($fieldLayout);
         }
 
-        return parent::beforeDelete();
+        return true;
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function afterRestore()
+    {
+        // Restore the field layout too
+        if (!Craft::$app->getFields()->restoreLayoutById($this->fieldLayoutId)) {
+            Craft::warning("Global set {$this->id} restored, but its field layout ({$this->fieldLayoutId}) was not.");
+        }
+
+        parent::afterRestore();
     }
 }
